requirements.txt
streamlit==1.29.0
langchain==0.0.352
langchain-community==0.0.38
langchain-huggingface==0.0.1
chromadb==0.4.22
python-dotenv==1.0.0
tiktoken==0.5.2
pypdf==4.0.1
docx2txt==0.8
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.1
accelerate==0.24.1
huggingface-hub==0.19.4
faiss-cpu==1.7.4
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9